feat: extracted cart to it's own component. Replaced toast with checkout button next to pagination
